Enable launch eloqkv cluster with in memory mode. (#258)

+++ src/cluster_mgr/src/cli/task/group/remove_group.rs
@@ -102,36 +102,37 @@ impl TaskGroup for RemoveTaskGroup {
         barrier.push(clean_tasks.len());
         executable.extend(clean_tasks);
         // remove keyspace in external cassandra/scylla/dynamo
-        let store = &cluster_config.deployment.storage_service;
-        match store.provider().unwrap() {
-            StorageProvider::Cassandra => {
-                let cass = store.cassandra.as_ref().unwrap();
-                if cass.external().is_some() {
-                    let host = cass.host.first().unwrap().clone();
-                    let task_id = TaskId {
-                        cmd: "remove".to_string(),
-                        task: "drop-keyspace".to_string(),
-                        host: "_local".to_string(),
-                    };
-                    let cql = format!(
-                        "DROP KEYSPACE {}",
-                        cluster_config.deployment.get_keyspace()?
-                    );
-                    let task =
-                        CassandraOpTask::new(task_id.clone(), host, cass.client_port()?, cql);
-                    let inst = TaskInstance {
-                        task_input: HashMap::default(),
-                        task: Box::new(task),
-                        task_host: TaskHost::Local,
-                    };
-                    barrier.push(1);
-                    executable.insert(task_id, inst);
+        if let Some(store) = &cluster_config.deployment.storage_service {
+            match store.provider().unwrap() {
+                StorageProvider::Cassandra => {
+                    let cass = store.cassandra.as_ref().unwrap();
+                    if cass.external().is_some() {
+                        let host = cass.host.first().unwrap().clone();
+                        let task_id = TaskId {
+                            cmd: "remove".to_string(),
+                            task: "drop-keyspace".to_string(),
+                            host: "_local".to_string(),
+                        };
+                        let cql = format!(
+                            "DROP KEYSPACE {}",
+                            cluster_config.deployment.get_keyspace()?
+                        );
+                        let task =
+                            CassandraOpTask::new(task_id.clone(), host, cass.client_port()?, cql);
+                        let inst = TaskInstance {
+                            task_input: HashMap::default(),
+                            task: Box::new(task),
+                            task_host: TaskHost::Local,
+                        };
+                        barrier.push(1);
+                        executable.insert(task_id, inst);
+                    }
                 }
+                StorageProvider::Dynamodb => {
+                    bail!("drop dynamodb keyspace is not implemented")
+                }
+                StorageProvider::Rocksdb => {}
             }
-            StorageProvider::Dynamodb => {
-                bail!("drop dynamodb keyspace is not implemented")
-            }
-            StorageProvider::Rocksdb => {}
         }
 
         let remove_backup_task = BackupTaskGroup
